requirements.txt
gdown==5.2.0
pandas==2.3.2
psycopg2-binary==2.9.10
sqlalchemy==1.4.36
sqlparse==0.5.3

